Output bwa_index files to tarball
First, it is necessary to copy the input FASTA (which gets mounted)
to the working directory so that output files will be created in the
working directory. To prevent having to pass each of these files as
inputs to the next task, I followed the reasoning of the ww-star module
and put them into a tarball.