Remove GHA to test vignettes
As vignettes are built out they need varied and complex input test data. It
is burdensome to download all test data needed for an entire workflow within
a GHA yaml script. This is partly because of the tools (e.g. bcftools) needed
to process the test data.

Instead, the new plan will be to test vignettes locally.